--- a/setutil_arm64.s
+++ b/setutil_arm64.s
@@ -53,11 +53,11 @@ TEXT ·union2by2(SB), NOSPLIT, $0-80
 	//Save the starting register of buffer.
 	MOVD buffer+48(FP), R5
 
-    // set1 is empty, just flush set2
+	// set1 is empty, just flush set2
 	CMP R0, R3
 	BEQ flush_right
 
-    // set2 is empty, just flush set1
+	// set2 is empty, just flush set1
 	CMP R1, R4
 	BEQ flush_left
 
@@ -69,64 +69,64 @@ TEXT ·union2by2(SB), NOSPLIT, $0-80
 	MOVHU.P 2(R1), R7
 loop:
 
-    CMP R6, R7
-    BEQ pop_both // R6 == R7
-    BLS pop_right // R6 > R7
+	CMP R6, R7
+	BEQ pop_both // R6 == R7
+	BLS pop_right // R6 > R7
 //pop_left: // R6 < R7
-    MOVHU.P R6, 2(R2)
-    CMP R0, R3
-    BEQ pop_then_flush_right
-    MOVHU.P 2(R0), R6
-    JMP loop
+	MOVHU.P R6, 2(R2)
+	CMP R0, R3
+	BEQ pop_then_flush_right
+	MOVHU.P 2(R0), R6
+	JMP loop
 pop_both:
-    MOVHU.P R6, 2(R2) //could also use R7, since they are equal
-    CMP R0, R3
-    BEQ flush_right
-    CMP R1, R4
-    BEQ flush_left
+	MOVHU.P R6, 2(R2) //could also use R7, since they are equal
+	CMP R0, R3
+	BEQ flush_right
+	CMP R1, R4
+	BEQ flush_left
 	MOVHU.P 2(R0), R6
 	MOVHU.P 2(R1), R7
 	JMP loop
 pop_right:
-    MOVHU.P R7, 2(R2)
-    CMP R1, R4
-    BEQ pop_then_flush_left
-    MOVHU.P 2(R1), R7
-    JMP loop
+	MOVHU.P R7, 2(R2)
+	CMP R1, R4
+	BEQ pop_then_flush_left
+	MOVHU.P 2(R1), R7
+	JMP loop
 
 pop_then_flush_right:
-    MOVHU.P R7, 2(R2)
+	MOVHU.P R7, 2(R2)
 flush_right:
 	MOVD R1, R0
 	MOVD R4, R3
 	JMP flush_left
 pop_then_flush_left:
-    MOVHU.P R6, 2(R2)
+	MOVHU.P R6, 2(R2)
 flush_left:
-    CMP R0, R3
-    BEQ return
-    //figure out how many bytes to slough off. Must be a multiple of two
-    SUB R0, R3, R4
-    ANDS $6, R4
-    BEQ long_flush //handles the 0 mod 8 case
-    SUBS $4, R4, R4 // since possible values are 2, 4, 6, this splits evenly
-    BLT pop_single  // exactly the 2 case
-    MOVW.P 4(R0), R6
-    MOVW.P R6, 4(R2)
-    BEQ long_flush // we're now aligned by 64 bits, as R4==4, otherwise 2 more
+	CMP R0, R3
+	BEQ return
+	//figure out how many bytes to slough off. Must be a multiple of two
+	SUB R0, R3, R4
+	ANDS $6, R4
+	BEQ long_flush //handles the 0 mod 8 case
+	SUBS $4, R4, R4 // since possible values are 2, 4, 6, this splits evenly
+	BLT pop_single  // exactly the 2 case
+	MOVW.P 4(R0), R6
+	MOVW.P R6, 4(R2)
+	BEQ long_flush // we're now aligned by 64 bits, as R4==4, otherwise 2 more
 pop_single:
-    MOVHU.P 2(R0), R6
-    MOVHU.P R6, 2(R2)
+	MOVHU.P 2(R0), R6
+	MOVHU.P R6, 2(R2)
 long_flush:
-    // at this point we know R3 - R0 is a multiple of 8.
-    CMP R0, R3
-    BEQ return
-    MOVD.P 8(R0), R6
-    MOVD.P R6, 8(R2)
-    JMP long_flush
+	// at this point we know R3 - R0 is a multiple of 8.
+	CMP R0, R3
+	BEQ return
+	MOVD.P 8(R0), R6
+	MOVD.P R6, 8(R2)
+	JMP long_flush
 return:
-    // number of shorts written is (R5 - R2) >> 1
-    SUB R5, R2
+	// number of shorts written is (R5 - R2) >> 1
+	SUB R5, R2
 	LSR $1, R2, R2
-    MOVD R2, size+72(FP)
-    RET
+	MOVD R2, size+72(FP)
+	RET