--- a/setutil_generic.go
+++ b/setutil_generic.go
@@ -1,4 +1,4 @@
-// !arm64 gccgo appengine
+// +build !arm64 gccgo appengine
 
 package roaring