This is an initial pass at implementing a setutils method (union2by2) in assembly, in particular in arm64 assembly. It speeds up union2by2 by 40-50% compared to the gc based code. 

In order to allow for differential compiling I created setutils_generic.go and setutils_arm64.go, following the pattern of serialization and popcount in this package.

I don't have the final `copy()` calls that the generic code does because I'm still new to golang assembly and couldn't figure it out. The solution I have is decently fast, as it will copy 64 bits at a time, but if someone wanted point me at the right way to do `CALL runtime.memmove()`, that'd be great.

```
benchstat --delta-test none -sort -delta  old.txt new.txt
name                                    old time/op    new time/op    delta
RepeatedSelfArrayUnion-16                 15.0ms ± 0%     6.9ms ± 0%  -54.38%
RepeatedGrowthArrayUnion-16               7.89ms ± 0%    3.71ms ± 0%  -52.98%
UnionInPlaceCopyOnWrite/4-4-3-10-16       55.9µs ± 0%    33.0µs ± 0%  -41.00%
UnionInPlaceCopyOnWrite/10-10-3-10-16     3.62ms ± 0%    2.14ms ± 0%  -40.88%
UnionInPlaceCopyOnWrite/7-7-3-10-16        447µs ± 0%     265µs ± 0%  -40.75%
UnionInPlaceCopyOnWrite/10-7-3-10-16       453µs ± 0%     271µs ± 0%  -40.24%
UnionInPlaceCopyOnWrite/7-4-3-10-16       56.8µs ± 0%    34.0µs ± 0%  -40.21%
UnionInPlaceCopyOnWrite/7-10-3-10-16       481µs ± 0%     292µs ± 0%  -39.37%
UnionInPlaceCopyOnWrite/4-7-3-10-16       61.1µs ± 0%    37.9µs ± 0%  -37.93%
UnionInPlaceCopyOnWrite/10-4-3-10-16      60.7µs ± 0%    38.5µs ± 0%  -36.56%
EvenIntervalArrayUnions-16                36.3µs ± 0%    23.1µs ± 0%  -36.42%
ArrayUnionThreshold/mostly-overlap-16     14.7µs ± 0%     9.7µs ± 0%  -34.02%
ArrayUnionThreshold/little-overlap-16     17.5µs ± 0%    11.8µs ± 0%  -32.59%
InPlaceArrayUnions-16                      649µs ± 0%     443µs ± 0%  -31.67%
UnionInPlaceCopyOnWrite/4-10-3-10-16      89.1µs ± 0%    65.4µs ± 0%  -26.66%
UnionInPlaceCopyOnWrite/4-4-10-10-16       210µs ± 0%     154µs ± 0%  -26.60%
UnionInPlaceCopyOnWrite/10-7-10-3-16       363µs ± 0%     267µs ± 0%  -26.54%
UnionInPlaceCopyOnWrite/4-4-10-3-16       43.9µs ± 0%    32.4µs ± 0%  -26.30%
UnionInPlaceCopyOnWrite/7-4-10-3-16       45.1µs ± 0%    33.2µs ± 0%  -26.29%
UnionInPlaceCopyOnWrite/7-4-10-10-16       211µs ± 0%     156µs ± 0%  -26.19%
UnionInPlaceCopyOnWrite/10-10-10-3-16     2.90ms ± 0%    2.15ms ± 0%  -25.70%
UnionInPlaceCopyOnWrite/10-10-10-10-16    13.6ms ± 0%    10.1ms ± 0%  -25.66%
UnionInPlaceCopyOnWrite/7-7-10-10-16      1.69ms ± 0%    1.26ms ± 0%  -25.50%
UnionInPlaceCopyOnWrite/10-7-10-10-16     1.70ms ± 0%    1.27ms ± 0%  -25.44%
UnionInPlaceCopyOnWrite/10-4-10-10-16      214µs ± 0%     160µs ± 0%  -25.29%
UnionInPlaceCopyOnWrite/4-7-10-10-16       215µs ± 0%     161µs ± 0%  -25.21%
UnionInPlaceCopyOnWrite/7-7-10-3-16        354µs ± 0%     265µs ± 0%  -25.11%
UnionInPlaceCopyOnWrite/7-10-10-3-16       392µs ± 0%     295µs ± 0%  -24.80%
UnionInPlaceCopyOnWrite/4-7-10-3-16       49.1µs ± 0%    37.0µs ± 0%  -24.67%
UnionInPlaceCopyOnWrite/7-10-10-10-16     1.71ms ± 0%    1.29ms ± 0%  -24.66%
UnionInPlaceCopyOnWrite/10-4-10-3-16      50.0µs ± 0%    37.8µs ± 0%  -24.31%
UnionInPlaceCopyOnWrite/4-10-10-10-16      243µs ± 0%     187µs ± 0%  -23.00%
AntagonisticArrayUnionsGrowth-16          50.7µs ± 0%    39.2µs ± 0%  -22.63%
ArrayUnionThreshold/no-overlap-16         32.0µs ± 0%    25.9µs ± 0%  -19.00%
UnionInPlaceCopyOnWrite/4-10-10-3-16      77.3µs ± 0%    65.2µs ± 0%  -15.65%
UnionInPlaceCopyOnWrite/7-7-3-3-16        27.5µs ± 0%    23.3µs ± 0%  -15.16%
UnionInPlaceCopyOnWrite/10-10-3-3-16       235µs ± 0%     200µs ± 0%  -14.70%
UnionInPlaceCopyOnWrite/7-4-3-3-16        3.95µs ± 0%    3.41µs ± 0%  -13.66%
UnionInPlaceCopyOnWrite/4-4-3-3-16        3.20µs ± 0%    2.77µs ± 0%  -13.55%
UnionInPlaceCopyOnWrite/10-7-3-3-16       32.9µs ± 0%    28.4µs ± 0%  -13.46%
UnionInPlaceCopyOnWrite/7-10-3-3-16       57.2µs ± 0%    52.2µs ± 0%   -8.79%
UnionInPlaceCopyOnWrite/4-7-3-3-16        7.47µs ± 0%    6.90µs ± 0%   -7.59%
UnionInPlaceCopyOnWrite/10-4-3-3-16       9.62µs ± 0%    9.04µs ± 0%   -6.03%
UnionInPlaceCopyOnWrite/4-10-3-3-16       37.0µs ± 0%    36.2µs ± 0%   -2.03%
UnionBitset-16                            2.88ms ± 0%    2.88ms ± 0%   -0.01%
UnionRoaring-16                            286µs ± 0%     289µs ± 0%   +0.97%
```