@lemire 2020-10-29
See https://github.com/RoaringBitmap/roaring/issues/289