@lemire 2020-10-29
Can you elaborate on why you think that memmove would help? I have not read your code in great details, but memmove is not magic, right? It might be more optimized than your code, but you are paying for the function call.... so....