@jacksonrnewhouse 2020-10-30
To initially familiarize myself with arm64 assembly I started by reimplementing copy() for two equal length short slices. The result is basically that tail code. It's 5x faster than iterating short by short, and 30% faster than converting to longs using unsafe and copying long by long. I also benchmarked copy() and it was over twice as fast as my code, hence seeming like magic. I thought about putting flush in neon, which might beat it out, but it also doesn't seem like a huge problem, as we're talking about the easy happy case.