@lemire 2020-10-30
Anyhow, I am happy to merge this if you think it is ready. Things are all green.