@jacksonrnewhouse 2020-10-30
I think merging this is very reasonable. I'm curious if we have anyone using this at scale on arm64 and could provide feedback on the performance of this and #286, as imagine both will move the needle considerably. I do plan on taking a stab at #288, as faster is better. Will also try to optimize at least 2by2localIntersectionCardinality.