@lemire 2020-10-30
Merging.